cleaned up PDFWorker process method

+++ pypln/workers/pdfconv_worker.py
@@ -33,24 +33,28 @@ class PDFConverterWorker(PushPullWorker):
             stdout, stderr = p.communicate(input=fi)
             # Also get file Metadata
             q = subprocess.Popen(['pdfinfo','-meta','-','-'],stdin=subprocess.PIPE,  stdout=subprocess.PIPE)
-            metadata, mdderr = p.communicate(input=fi)
+            metadata, mderr = p.communicate(input=fi)
             metadata = self.parse_metadata(metadata)
+            if not (stdout and metadata):
+                msgout = {'fail':1}
+            else:
+                if not (stderr or mderr):
+                    msgout = {'filename':pdf.filename,'text':stdout.strip(),'file_metadata':metadata,
+                              'database':msg['database'],'collection':msg['collection']}
+                else:
+                    msgout = {'fail':1}
+                    p.kill()
+                    q.kill()
         except OSError:
             p.kill()
             q.kill()
-            self.sender.send_json({'fail':1})
-            return
+            msgout = {'fail':1}
+        finally:
+            self.sender.send_json(msgout)
+        return
         # Assumes encoding is utf-8 which is not guaranteed. Although pdftotext attempts to convert to utf-8 it may not work.
-        if not stdout:
-            self.sender.send_json({'fail':1})
-        else:
-            if not stderr:
-                msgout = {'filename':pdf.filename,'text':stdout.strip(),'file_metadata':metadata,
-                          'database':msg['database'],'collection':msg['collection']} #stripping to remove non-printing characters
-                self.sender.send_json(msgout)
-            else:
-                self.sender.send_json({'fail':1})
-                p.kill()
+
+
 
     def parse_metadata(self,md):
         """
